Refactoring: Merged parts of User and Authentication feature into new Account feature
