+++ src/app/store/middleware.ts
@@ -1,4 +1,4 @@
 import { requestHandler } from 'features/requests'
-import { authHandler } from 'features/authentication'
+import { authHandler } from 'features/account'
 
 export default [requestHandler, authHandler]
