+++ src/features/account/components/hooks.ts
@@ -1,0 +1,4 @@
+import { loadIfNotLoaded } from 'features/requests'
+import { loadAccount } from 'features/account/store'
+
+export const useAccountData = () => loadIfNotLoaded(loadAccount)
