+++ src/features/account/components/index.ts
@@ -1,0 +1,6 @@
+import LoginForm from './login-form/LoginForm'
+export { LoginForm }
+
+export { AccountInfo } from './account-info/AccountInfo'
+export { AccountEditForm, RegistrationForm } from './account-form'
+export { useAccountData } from './hooks'
