+++ src/features/account/store/types.ts
@@ -1,0 +1,13 @@
+import { User } from 'features/users'
+
+export interface AccountState {
+  account: Account | null
+  token: string | null
+}
+
+export type Account = User
+
+export interface UserCredentials {
+  email: string
+  password: string
+}
