--- src/features/authentication/components/index.ts
@@ -1,3 +1,0 @@
-import LoginForm from 'features/authentication/components/LoginForm'
-
-export { LoginForm }
