--- src/features/authentication/index.ts
@@ -1,3 +1,0 @@
-export { LoginForm } from './components'
-export { withAuthGuard } from './hocs'
-export { isLoggedIn, authHandler, logout } from './store'
