--- src/features/authentication/store/index.ts
@@ -1,4 +1,0 @@
-export { login, logout } from './actions'
-export { isLoggedIn } from './selectors'
-export { AuthState, UserCredentials } from './types'
-export { authHandler } from './middleware'
