--- src/features/authentication/store/types.ts
@@ -1,8 +1,0 @@
-export interface AuthState {
-  isLoggedIn: boolean
-}
-
-export interface UserCredentials {
-  email: string
-  password: string
-}
