+++ src/features/users/components/hooks.ts
@@ -1,5 +1,4 @@
 import { loadIfNotLoaded } from 'features/requests'
-import { loadAccount, loadBoards } from 'features/users/store'
+import { loadBoards } from 'features/users/store'
 
-export const useAccountData = () => loadIfNotLoaded(loadAccount)
 export const useBoardData = () => loadIfNotLoaded(loadBoards)
