+++ src/features/users/components/index.ts
@@ -1,4 +1,2 @@
-export { AccountInfo } from './account-info/AccountInfo'
 export { BoardList } from './board/BoardList'
-export { AccountEditForm, RegistrationForm } from './account-form'
-export { useAccountData } from './hooks'
+export { UserProfile } from './user/UserProfile'
