+++ src/features/users/components/user/ModeratorIcon.tsx
