+++ src/features/users/index.ts
@@ -1,2 +1,2 @@
-export { AccountInfo, RegistrationForm, AccountEditForm } from './components'
-export { usersReducer, UsersState } from './store'
+export { BoardList } from './components'
+export { usersReducer, UsersState, User, addBoardsFromResponse } from './store'
