+++ src/features/users/store/actions.ts
@@ -1,10 +1,7 @@
-export { ACCOUNT_LOAD_REQUEST_ID, ACCOUNT_SET, loadAccount, patchAccount } from './actions.account'
-export { BOARDS_REQUEST_ID, BOARDS_ADD, loadBoards } from './actions.boards'
+export { BOARDS_REQUEST_ID, BOARDS_ADD, loadBoards, addBoardsFromResponse } from './actions.boards'
 export { USERS_ADD } from './actions.users'
-export { REGISTRATION_REQUEST_ID, register } from './actions.registration'
 
-import { AccountActions } from './actions.account'
 import { BoardActions } from './actions.boards'
 import { UserActions } from './actions.users'
 
-export type UserActionTypes = AccountActions | BoardActions | UserActions
+export type UserActionTypes = BoardActions | UserActions
