+++ src/features/users/store/index.ts
@@ -1,5 +1,5 @@
 export { usersReducer } from './reducer'
-export { UsersState, Account, Board, User } from './types'
-export { AccountResponse, BoardResponse, UserResponse, RegistrationRequest } from './types.api'
-export { loadAccount, loadBoards, register, patchAccount } from './actions'
-export { registrationSucceeded, getAccount, getBoard, getUser, getUsersByIds } from './selectors'
+export { UsersState, Board, User } from './types'
+export { BoardResponse, UserResponse } from './types.api'
+export { loadBoards, addBoardsFromResponse } from './actions'
+export { getBoard, getUser, getUsersByIds } from './selectors'
