+++ src/features/users/store/selectors.ts
@@ -1,10 +1,7 @@
 import { ApplicationState } from 'app/store/reducers'
 import { UsersState, User, Board } from 'features/users/store/types'
-import { isRequestDone } from 'features/requests/store'
-import { REGISTRATION_REQUEST_ID } from 'features/users/store/actions'
 
 const getSlice = (state: ApplicationState): UsersState => state.users
-export const getAccount = (state: ApplicationState): UsersState['account'] => getSlice(state).account
 
 const getBoards = (state: ApplicationState): UsersState['boards'] => getSlice(state).boards
 export const getBoard = (id: Board['id']) => (state: ApplicationState): Board | undefined => getBoards(state)[id]
@@ -17,5 +14,3 @@ export const getUser = (id: User['id']) => (state: ApplicationState): User | und
 export const getUsersByIds = (ids: User['id'][]) => (state: ApplicationState): User[] => [
   ...ids.map(id => getUsers(state)[id]).filter(user => user !== undefined),
 ]
-
-export const registrationSucceeded = isRequestDone(REGISTRATION_REQUEST_ID)
