+++ src/features/users/store/types.api.ts
@@ -1,18 +1,5 @@
 import { CountryResponse } from 'features/countries'
 
-export interface AccountResponse {
-  id: number
-  email: string
-  country: CountryResponse
-  firstname: string
-  lastname: string
-  currentCountry: CountryResponse
-  boards: BoardResponse[]
-  isModerator: boolean
-  created: string
-  updated: string
-}
-
 export interface BoardResponse {
   id: number
   country: CountryResponse
@@ -38,20 +25,4 @@ export interface UserResponse {
   currentCountry: CountryResponse
   groups: [] //?
   userPermissions: [] //?
-}
-
-export interface RegistrationRequest {
-  firstname: string
-  lastname: string
-  email: string
-  password: string
-  country: number
-}
-
-export interface AccountPatchRequest {
-  firstname?: string
-  lastname?: string
-  email?: string
-  password?: string
-  country?: number
 }
