+++ src/features/users/store/types.ts
@@ -1,13 +1,8 @@
 import { Country } from 'features/countries'
 
 export type UsersState = {
-  account: Account | {}
   boards: { [id in Board['id']]: Board }
   users: { [id in User['id']]: User }
-}
-
-export interface Account extends User {
-  boards: Board['id'][]
 }
 
 export interface Board {
@@ -27,4 +22,5 @@ export interface User {
   lastName: string
   country: Country['id']
   isModerator: boolean
+  boards?: Board['id'][]
 }
