+++ src/pages/Login.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { LoginForm } from 'features/authentication'
+import { LoginForm } from 'features/account'
 
 const Login = () => <LoginForm />
